requirements.txt
eth-brownie==1.18.1
pytest==6.2.5